+++ stories/index.stories.js
@@ -12,6 +12,7 @@ import StInput from '../src/components/input/index.vue';
 import StPopover from '../src/components/popover/index.vue';
 import StRadio from '../src/components/radio/index.vue';
 import StRow from '../src/components/row/index.vue';
+import StScrollbar from '../src/components/scrollbar/index.vue';
 import StSwitch from '../src/components/switch/index.vue';
 import StTextarea from '../src/components/textarea/index.vue';
 
@@ -26,5 +27,6 @@ Vue.component('st-input', StInput);
 Vue.component('st-popover', StPopover);
 Vue.component('st-radio', StRadio);
 Vue.component('st-row', StRow);
+Vue.component('st-scrollbar', StScrollbar);
 Vue.component('st-switch', StSwitch);
 Vue.component('st-textarea', StTextarea);