+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## 0.2.0
 
+### Features
+* Added `Radio` component to `Offers` theme
+
 ### Maintenance
 * Fixed `Offers` theme for `Button` component
-* Added `Radio` component to `Offers` theme
 * Fixed `Checkbox` value type 
+* Added components generator. If you are about to make a new component just run `npm run generator:component` and follow the instructions.
 
 ### Features
 * Added `Scrollbar` component