+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 * Added `Radio` component to `Offers` theme
 * Added new component: `Notifications`
 * Added new component: `Scrollbar`
+* Added new component: `Pagination`
 
 ### Maintenance
 * Fixed `Offers` theme for `Button` component
 * Fixed `Checkbox` value type 
 * Added components generator. If you are about to make a new component just run `npm run generator:component` and follow the instructions.
+* Prepared for publishing